# docs/contracts/openapi.yaml
openapi: 3.1.1
info:
  title: Modelry Community Core HTTP API
  version: 0.1.0
  description: >-
    Canonical V0.1 HTTP contract for the Modelry Control Plane and Application API.
    Route groups keep the two product planes explicit and share request and error semantics.
servers:
  - url: /
security:
  - OwnerSession: []
  - ServiceAccountAPIKey: []
...
paths:
...
  '/admin/api/v1/runtime/status':
    get:
      tags: [Runtime]
      operationId: getRuntimeStatus
      summary: Read the public Runtime readiness snapshot
      security:
        - OwnerSession: []
        - ServiceAccountAPIKey: []
        - {}
      responses:
        '200':
          description: Successful response.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeStatusResponse'
        '4XX':
          $ref: '#/components/responses/StructuredError'
        '5XX':
          $ref: '#/components/responses/StructuredError'
        default:
          $ref: '#/components/responses/StructuredError'
  '/admin/api/v1/storage/status':
    get:
      tags: [Runtime]
      operationId: getStorageStatus
      summary: Read the public SQLite and Local Storage health snapshot
      security:
        - OwnerSession: []
        - ServiceAccountAPIKey: []
        - {}
      responses:
        '200':
          description: Successful response.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StorageStatusResponse'
        '4XX':
          $ref: '#/components/responses/StructuredError'
        '5XX':
          $ref: '#/components/responses/StructuredError'
        default:
          $ref: '#/components/responses/StructuredError'
...
  '/admin/api/v1/requests':
    get:
      tags: [Requests]
      operationId: listRequestRecords
      summary: List redacted Application RequestRecords
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Search'
        - $ref: '#/components/parameters/Filter'
        - $ref: '#/components/parameters/Sort'
      responses:
        '200':
          description: Successful response.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestRecordListResponse'
        '4XX':
          $ref: '#/components/responses/StructuredError'
        '5XX':
          $ref: '#/components/responses/StructuredError'
        default:
          $ref: '#/components/responses/StructuredError'
  '/admin/api/v1/requests/{requestId}':
    parameters:
      - $ref: '#/components/parameters/RequestId'
    get:
      tags: [Requests]
      operationId: getRequestRecord
      summary: Read the Request Detail for a canonical requestId
      parameters:
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Successful response.
          headers:
            X-Request-Id:
              $ref: '#/components/headers/RequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestRecordResponse'
        '4XX':
          $ref: '#/components/responses/StructuredError'
        '5XX':
          $ref: '#/components/responses/StructuredError'
        default:
          $ref: '#/components/responses/StructuredError'
...
components:
  securitySchemes:
    OwnerSession:
      type: apiKey
      in: cookie
      name: modelry_admin_session
      description: HttpOnly Owner session cookie created by the Control Plane login operation.
    ServiceAccountAPIKey:
      type: http
      scheme: bearer
      description: One-time-revealed Service Account API Key; authorization still checks Permission.
    ApplicationSession:
      type: http
      scheme: bearer
      description: Opaque, server-revocable Application Session Token.
  headers:
    RequestId:
      description: Canonical server-generated request ID for this HTTP request.
      schema:
        $ref: '#/components/schemas/RequestId'
...
  parameters:
...
    RequestId:
      name: requestId
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/RequestId'
...
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
    Search:
      name: search
      in: query
      required: false
      schema:
        type: string
    Filter:
      name: filter
      in: query
      required: false
      description: >-
        One filter only, in the form `fieldName operator JSON-scalar`. Operators are eq, ne, gt, gte, lt, lte, and contains. The field must be in the Applied Model; contains is text-only, ordered comparisons support text, number, and dateTime, and null supports only eq/ne. Arbitrary SQL is forbidden.
      schema:
        type: string
    Sort:
      name: sort
      in: query
      required: false
      description: Comma-separated field keys with optional asc or desc; no SQL expressions.
      schema:
        type: string
...
  responses:
    StructuredError:
      description: Structured error response with the canonical request ID.
      headers:
        X-Request-Id:
          $ref: '#/components/headers/RequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    Id:
      type: string
      minLength: 1
      description: Opaque stable domain identity; not a SQLite row identifier.
    RequestId:
      type: string
      pattern: '^req_[A-Za-z0-9_-]{8,}$'
      description: Canonical server-generated ID shared by response header, Error and RequestRecord.
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
      additionalProperties: false
    ApiError:
      type: object
      description: For anonymous Runtime or Storage diagnostics, no error field may reveal an absolute Project Root path or secret.
      required: [code, message, details, requestId]
      properties:
        code:
          type: string
          enum: [INVALID_ARGUMENT, UNAUTHENTICATED, FORBIDDEN, NOT_FOUND, CONFLICT, BOOTSTRAP_CLOSED, CHANGE_CONFIRMATION_REQUIRED, PAYLOAD_TOO_LARGE, UNSUPPORTED_MEDIA_TYPE, VALIDATION_FAILED, REGISTRATION_DISABLED, RATE_LIMITED, STORAGE_UNAVAILABLE, STORAGE_BUSY, RUNTIME_NOT_READY, INTERNAL_ERROR]
        message:
          type: string
          description: Human-readable message without stack traces, SQL, credentials, sensitive body content, or anonymous diagnostic paths.
        details:
          $ref: '#/components/schemas/ErrorDetails'
        hint:
          type: string
          description: Optional safe action; anonymous diagnostic hints omit absolute Project Root paths and secrets.
        requestId:
          $ref: '#/components/schemas/RequestId'
      additionalProperties: false
    ErrorDetails:
      type: object
      description: Structured details; anonymous diagnostic error details must omit absolute Project Root paths and secrets.
      properties:
        violations:
          type: array
          items:
            $ref: '#/components/schemas/Violation'
      additionalProperties: true
    Violation:
      type: object
      required: [path, code, message]
      properties:
        path:
          type: string
        code:
          type: string
        message:
          type: string
      additionalProperties: true
...
    RuntimeStatusResponse:
      type: object
      description: Safe readiness snapshot; never includes the Project Root absolute path or secrets.
      required: [state, observedAt, database, localStorage]
      properties:
        state:
          type: string
          enum: [starting, ready, degraded, unavailable, stopping]
        observedAt:
          type: string
          format: date-time
        database:
          $ref: '#/components/schemas/Health'
        localStorage:
          $ref: '#/components/schemas/Health'
        projectSource:
          type: string
          description: Identifies flag, environment, or working-directory selection without returning its value or any path.
        version:
          type: string
      additionalProperties: false
    StorageStatusResponse:
      type: object
      description: Safe storage snapshot; the localStorage path is omitted unless the request has valid diagnostic authorization.
      required: [database, localStorage]
      properties:
        database:
          $ref: '#/components/schemas/Health'
        localStorage:
          $ref: '#/components/schemas/LocalStorageHealth'
      additionalProperties: false
    LocalStorageHealth:
      type: object
      description: Contains the absolute Local Storage path only for a request authenticated and authorized to read diagnostics.
      required: [state, provider]
      properties:
        state:
          type: string
          enum: [ready, degraded, unavailable, unknown]
        message:
          type: string
          description: Safe diagnostic text; must not include an absolute Project Root path or secret.
        hint:
          type: string
          description: Safe actionable guidance; must not include an absolute Project Root path or secret.
        provider:
          const: Local
        path:
          type: string
          description: Omitted for anonymous requests; returned only when Owner or Service Account Permission authorizes diagnostic path access.
      additionalProperties: false
    Health:
      type: object
      description: Safe diagnostic health; message and hint never include an absolute Project Root path or secret.
      required: [state]
      properties:
        state:
          type: string
          enum: [ready, degraded, unavailable, unknown]
        message:
          type: string
        hint:
          type: string
      additionalProperties: false
...
    RequestRecord:
      type: object
      required: [requestId, time, endpoint, method, status, durationMs]
      properties:
        requestId:
          $ref: '#/components/schemas/RequestId'
        time:
          type: string
          format: date-time
        collectionId:
          $ref: '#/components/schemas/Id'
        endpoint:
          type: string
        method:
          type: string
        status:
          type: integer
          minimum: 100
          maximum: 599
        durationMs:
          type: integer
          minimum: 0
        responseSizeBytes:
          type: integer
          format: int64
          minimum: 0
          description: Response body size in bytes. The body itself is never stored. Omitted for RequestRecords created before this measurement was introduced.
        authenticationOutcome:
          type: string
        authorizationOutcome:
          type: string
        errorCode:
          type: string
      additionalProperties: false
    RequestRecordResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/RequestRecord'
    RequestRecordListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/RequestRecord'
        nextCursor:
          type: string
...